+++ skyperious/conf.py
@@ -10,7 +10,7 @@ Released under the MIT License.
 
 @author      Erki Suurjaak
 @created     26.11.2011
-@modified    16.02.2021
+@modified    22.02.2021
 ------------------------------------------------------------------------------
 """
 from ConfigParser import RawConfigParser
@@ -24,8 +24,8 @@ import appdirs
 
 """Program title, version number and version date."""
 Title = "Skyperious"
-Version = "4.5.1.dev7"
-VersionDate = "16.02.2021"
+Version = "4.5.1.dev8"
+VersionDate = "22.02.2021"
 
 if getattr(sys, "frozen", False):
     # Running as a pyinstaller executable