+++ CHANGELOG.md
@@ -2,11 +2,12 @@ CHANGELOG
 =========
 
 
-4.5.1, 2021-02-14
+4.5.1, 2021-02-22
 -----------------
 - show <pre>-tag content in chat reader;
 - show @-taggings in chat reader and style them bold in reader and HTML;
 - ask for confirmation when exporting HTML with media in subfolder, but login unavailable for downloading media (issue #93).
+- fixed merge not showing correct diff in individual chats view;
 - fixed duplicates in syncing bot contacts and messages from live (issue #93);
 - fixed not retaining the order of command-line FILE arguments (issue #93);
 - add detected media type extension to shared media exported in HTML if lacking one (issue #93).